requirements.txt
flask==3.0.3
groq==0.9.0
PyPDF2==3.0.1
python-docx==1.1.2
werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv
httpx==0.27.0